+++ lib/containers/protocols/flattenable.ex
@@ -1,0 +1,5 @@
+defprotocol Containers.Flattenable do
+  @spec flatten(Containers.flattenable)
+  :: Containers.flattenable
+  def flatten(flattable)
+end